Reuse DFF initial value when input is 'x
This helps to enforce the initial value during t0 on the very first
clock tick. Output of DFFs is typically 'x at this time, because it gets
set with a delay. DFFs that get the output of another DFF would not be
able to have their initial value, because it would be immediately
overwritten with a random value because of the incoming 'x.