Fix bit order of saved obj 6 x
The data for sprite (also know as objects) is saved in registers that go
from 0 to 9. When saving the X position in the register 6, the bit 6 and
4 were swapped.

This caused the test mooneye-teste-suite/acceptance/ppu/intr_2_mode0_timing_sprites.s
to fail due to wrong timing caused by the sprite X position being
incorrect.